+++ Assets/Scripts/Structs/AttackDetails.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 public struct AttackDetails
 {
     public Vector2 position;
-    public float damageAmount;
+    public int damageAmount;
     public float stunDamageAmount;
 }
